Removed logging for gui and added rudp dependency in cmakelist includes.

+++ src/maidsafe/lifestuff/detail/user_storage.cc
@@ -1152,7 +1152,7 @@ int UserStorage::InformContactsOperation(InboxItemType item_type,
     }
   }
 
-  return aggregate == contacts.size() ? aggregate : 0;
+  return aggregate == static_cast<int>(contacts.size()) ? aggregate : 0;
 }
 
 int UserStorage::InformContacts(InboxItemType item_type,
@@ -1208,7 +1208,7 @@ int UserStorage::InformContacts(InboxItemType item_type,
         contacts_results->insert(std::make_pair(it->first, result));
     }
   }
-  return aggregate == contacts.size() ? aggregate : 0;
+  return aggregate == static_cast<int>(contacts.size()) ? aggregate : 0;
 }
 
 std::string UserStorage::ConstructFile(const std::string &serialised_data_map) {
